Docs: Replace version by 'current' in changes.rst
The file 'Documentation/process/changes.rst' states the listed
requirements are for the 4.x kernel version. However, there are
requirements updated for the 5.x version, as there might be in other
future versions. This patch updates it to 'current kernel  version' so
the document won't be outdated in future kernel versions.

Signed-off-by: Bruno Moreira-Guedes <codeagain@codeagain.dev>